Postavljene scheme entitetima, model sa svim elementima

+++ src/main/java/hr/tvz/matkovic/service/ChemicalElementService.java
@@ -11,7 +11,7 @@ public interface ChemicalElementService {
 
     List<ChemicalElement> findAll();
 
-    ChemicalElement  findById(Long id);
+    ChemicalElement  findById(Integer id);
 
     ChemicalElement findByatomic_number(Integer atomicNumber);
 }
